fixed a lot of bugs

+++ src/Armor.java
@@ -1,7 +1,7 @@
 
 public class Armor extends Item{
-	int bonusHealth;
-	Armor(String name, int bonusHealth, boolean isActive) {
+	long bonusHealth;
+	Armor(String name, long bonusHealth, boolean isActive) {
 		super(name,isActive);
 		this.bonusHealth=bonusHealth;
 	}
